fixed dumb issues, reworked to use t_dir instead of DIR cuz opaque type

+++ src/dirent/ft_posix_getdents.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_posix_getdents.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lgoddijn <lgoddijn@student.codam.nl>       +#+  +:+       +#+        */
+/*   By: lgoddijn <lgoddijn@student.codam.nl >      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/16 00:47:34 by lgoddijn          #+#    #+#             */
-/*   Updated: 2024/10/16 00:52:29 by lgoddijn         ###   ########.fr       */
+/*   Updated: 2024/10/16 15:43:43 by lgoddijn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,5 +18,5 @@ ssize_t	ft_posix_getdents(int fd, void *buf, size_t len, int flags)
 		return (__syscall_r(-EOPNOTSUPP));
 	if (len > INT_MAX)
 		len = INT_MAX;
-	return (__syscall3(__NR_getdents, fd, buf, len));
+	return (__syscall3(__NR_getdents, fd, (long)buf, len));
 }
